+++ CHANGELOG.md
@@ -3,6 +3,24 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.11] - 2025-03-06
+
+### Added
+- Added Aimtell
+- Added Constant Contact
+- Added Font Awesome
+- Added Hotjar
+- Added Iterable
+- Added Klaviyo
+- Added Lightbox JS
+- Added Northbeam
+- Added Podscribe
+- Added Refersion
+- Added Swiftype
+- Added Usercentrics Cookiebot
+- Added Wisernotify
+- Added Wistia
+
 ## [1.0.10] - 2025-03-06
 
 ### Added